backend/requirements.txt
yt-dlp>=2024.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0
aiohttp>=3.9.0
apscheduler>=3.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.7
pydantic>=2.0
pydantic-settings>=2.0
websockets>=12.0
pysrt>=1.1.2
webvtt-py>=0.5.0
Pillow>=10.0
